# PI: Making pypdf as fast as pdfrw