--- a/pypdf/_cmap.py
+++ b/pypdf/_cmap.py
@@ -4,7 +4,7 @@ from math import ceil
 from typing import Any, Dict, List, Tuple, Union, cast
 
 from ._codecs import adobe_glyphs, charset_encoding
-from ._utils import logger_warning
+from ._utils import b_, logger_warning
 from .errors import PdfReadWarning
 from .generic import DecodedStreamObject, DictionaryObject, StreamObject
 
@@ -250,7 +250,7 @@ def prepare_cm(ft: DictionaryObject) -> bytes:
     tu = ft["/ToUnicode"]
     cm: bytes
     if isinstance(tu, StreamObject):
-        cm = cast(DecodedStreamObject, ft["/ToUnicode"]).get_data()
+        cm = b_(cast(DecodedStreamObject, ft["/ToUnicode"]).get_data())
     elif isinstance(tu, str) and tu.startswith("/Identity"):
         # the full range 0000-FFFF will be processed
         cm = b"beginbfrange\n<0000> <0001> <0000>\nendbfrange"