This is a follow-up of https://github.com/sarnold/pdfrw/issues/15#issuecomment-1676471854

Using the code from this branch, [pypdf_watermarking](https://github.com/py-pdf/benchmarks/blob/main/pdf_benchmark/library_code.py#L45) is as fast as `pdfrw_watermarking()`.

`pypdf` was made faster by avoiding unnecessarily calls to `__parse_content_stream()`.
In order to achieve that, some calls to `PageObject._push_pop_gs()` were removed in `PageObject._merge_page()` and `PageObject._merge_page_writer()`