@Lucas-C 2023-08-13
I'm surprised by the low number of failing tests:
`============= 7 failed, 773 passed, 4 xfailed in 449.28s (0:07:29) =============` (Python 3.11)

This is encouraging! 😄