@Lucas-C 2023-08-14
There are only 4 remaining tests that fail. They can be executed with:
`pytest -vv tests/test_workflows.py tests/test_writer.py -k 'test_extra_test_iss1541 or test_remove_images[side-by-side-subfig.pdf] or test_iss1601 or test_new_removes'`