@MartinThoma 2023-08-14
I've just merged the comment improvements into `main` and updated the branch so that those don't distract from the important changes.

I also just ran https://github.com/py-pdf/benchmarks to see the impact:

* Text extraction speed: 2.7s -> 2.5s (could be random)
* Image extraction speed: 3.3s -> 3.0s (could be random)
* Watermarking speed: 11.7s -> 0.5s :exclamation: :exploding_head: :astonished: 
* Watermarking file size: Decreased as well :thinking: