@pubpub-zz 2023-08-14
@Lucas-C 
I've proposed PR #1854 that allows to modify EncodedStreamData : Can't you just use this capability to directly add the waterwaterking on the "raw" encoded stream without converting it into ContentStream ?