@Lucas-C 2023-08-14
> I've proposed PR #1854 that allows to modify EncodedStreamData : Can't you just use this capability to directly add the waterwaterking on the "raw" encoded stream without converting it into ContentStream ?

Interesting.
I don't quite follow what "conversion into ContentStream" you are refering to...
You mean in `PageObject._merge_page()`?

> I also just ran [py-pdf/benchmarks](https://github.com/py-pdf/benchmarks) to see the impact:
> Watermarking speed: 11.7s → 0.5s ❗ 🤯 😲

Yes, that's promising!

By the way, I was wondering: would you be open to add non-regression unit tests focused on performances @MartinThoma?
We have a couple of them in `fpdf2`:
* https://github.com/py-pdf/fpdf2/blob/master/test/test_perfs.py#L10
* https://github.com/py-pdf/fpdf2/blob/master/test/fonts/test_charmap.py#L40

That have some drawbacks (those tests will fail when executed on slow computers),
but they are a good way to ensure that the code performances stay good over time,
for some a few given scenarios.