@MartinThoma 2023-08-15
@pubpub-zz I would appreciate some help here :-) 

## test_extra_test_iss1541

This test fails:

```
@pytest.mark.enable_socket()
def test_extra_test_iss1541():
    url = "https://github.com/py-pdf/pypdf/files/10418158/tst_iss1541.pdf"
    name = "tst_iss1541.pdf"
    data = BytesIO(get_data_from_url(url, name=name))
    reader = PdfReader(data, strict=False)
    reader.pages[0].extract_text()

    cs = ContentStream(reader.pages[0]["/Contents"], None, None)
    cs.operations.insert(-1, ([], b"EMC"))
    stream = BytesIO()
    cs.write_to_stream(stream)
    stream.seek(0)
    ContentStream(read_object(stream, None, None), None, None).operations

    cs = ContentStream(reader.pages[0]["/Contents"], None, None)
    cs.operations.insert(-1, ([], b"E!C"))
    stream = BytesIO()
    cs.write_to_stream(stream)
    stream.seek(0)
    with pytest.raises(PdfReadError) as exc:
        ContentStream(read_object(stream, None, None), None, None).operations
    assert exc.value.args[0] == "Unexpected end of stream"
```

No exception is thrown. I don't know if we would want an exception to be thrown there or if it's actually good that we no longer do it.

## test_writer.py::test_remove_images

```
    @pytest.mark.parametrize(
        "input_path",
        [
            "side-by-side-subfig.pdf",
            "reportlab-inline-image.pdf",
        ],
    )
    def test_remove_images(pdf_file_path, input_path):
        pdf_path = RESOURCE_ROOT / input_path
    
        reader = PdfReader(pdf_path)
        writer = PdfWriter()
    
        page = reader.pages[0]
        writer.insert_page(page, 0)
        writer.remove_images()
    
        # finally, write "output" to pypdf-output.pdf
        with open(pdf_file_path, "wb") as output_stream:
            writer.write(output_stream)
    
        with open(pdf_file_path, "rb") as input_stream:
            reader = PdfReader(input_stream)
            if input_path == "side-by-side-subfig.pdf":
                extracted_text = reader.pages[0].extract_text()
>               assert "Lorem ipsum dolor sit amet" in extracted_text
E               AssertionError: assert 'Lorem ipsum dolor sit amet' in ''
```

That is for sure not good.

## test_writer.py::test_iss1601

```
>       assert (
            ContentStream(reader.pages[0].get_contents(), reader).get_data()
            in page_1.get_contents().get_data()
        )

AssertionError: assert 
b'q\nq 1 0 0 1 0 0 cm /Xi0 Do Q\nQ\n\n1 0 0 1 0 0 cm  BT /F1 12 Tf 14.4 TL ET\nq\n1 0 0 1 34.24252 90.54567 cm\nq\n1 0...8 cm\nq\n0 0 0 rg\nBT 1 0 0 1 0 2.7 Tm 90.72463 0 Td /F2+0 18 Tf 20.7 TL (R M) Tj T* -90.72463 0 Td ET\nQ\nQ\nQ\n \n\n' in
b'q\n1 0.0 0.0 1 0.0 0.0 cm\n0.0 0.0 255.11811 153.07087000000001 re\nW\nn\nq\nq\n1 0 0 1 0 0 cm\n/Xi0 Do\nQ\nQ\n1 0 0...4630000000005 0 Td\n/F2+0 18 Tf\n20.699999999999999 TL\n(R\\040M) Tj\nT*\n-90.724630000000005 0 Td\nET\nQ\nQ\nQ\nQ\n\n'
```

The left-hand side starts with `q\nq 1 0`, but the right-hand side is `q\n1 0.0`. S left has an additional `Q` and the zero floating point formatting is different. This might be related to the file size.


## test_writer.py::test_new_removes

```
>       assert b"/Fm0 Do" in bb
E       assert b'/Fm0 Do' in b'%PDF-1.3\n%\xe2\xe3\xcf\xd3\n1 0 obj\n<<\n/Type /Catalog\n/Pages 2 0 R\n/Outlines 46 0 R\n>>\nendobj\n2 0 obj\n<<\n/Type /Pages\n/Count 3\n/Kids [ 3 0 R 42 0 R 44 0 R ]\n>>\nendobj\n3 0 obj\n<<\n/Type /Page\n/MediaBox [ 0 0 595.3039 841.88980000000004 ]\n/Resources 4 0 R\n/Rotate 0\n/Contents 48 0 R\n/Group <<\n/CS /DeviceRGB\n/I true\n/S /Transparency\n>>\n/Parent 2 0 R\n/Annots [ 29 0 R ]\n>>\nendobj\n4 0 obj\n<<\n/Font 5 0 R\n/XObject <<\n/Fm0 18 0 R\n/Fm1 24 0 R\n/Im0 25 0 R\n/Im4 26 0 R\n>>\n/ProcSet [ /ImageB /ImageC /ImageI /PDF /Text ]\n/ExtGState <<\n/GS0 27 0 R\n/GS1 21 0 R\n>>\n/ColorSpace <<\n/CS0 19 0 R\n>>\n>>\nendobj\n5 0 obj\n<<\n/F1 6 0 R\n/F2 10 0 R\n/TT0 14 0 R\n>>\nendobj\n6 0 obj\n<<\n/Type /Font\n/Subtype /TrueType\n/BaseFont /BAAAAA+LiberationSans-Bold\n/FirstChar 0\n/LastChar 51\n/Widths [ 750 666 666 722 666 777 777 722 722 666 610 722 610 277 277 722 610 389 610 556 610 556 333 277 333 556 556 556 556 556 277 889 556 277 722 277 777 556 610 500 333 610 556 610 610 610 722 610 610 833 333 666 ]\n/FontDescriptor 7 0 R\n/ToUnicode 9 0 R\n>>\nendobj\n7 0 obj\n<<\n/Type /FontDescriptor\n/FontName /BAAAAA+LiberationSans-Bold\n/Flags 4\n/FontBBox [ -481 -376 1303 1033 ]\n/ItalicAngle 0\n/Ascent 905\n/Descent -211\n/CapHeight 1033\n/StemV 80\n/FontFile2 8 0 R\n>>\nendobj\n8 0 obj\n<<\n/Filter /FlateDecode\n/Length1 17136\n/Length 
```