@MartinThoma 2023-08-15
> I've proposed PR https://github.com/py-pdf/pypdf/pull/1854 that allows to modify EncodedStreamData : Can't you just use this capability to directly add the waterwaterking on the "raw" encoded stream without converting it into ContentStream ?

Can you make an example PR?

I think this PR is interesting for three reasons (1) Huge performance improvement in watermarking (2) File size reductions (3) test_extra_test_iss1541 no longer throws an exception ... I'm not sure if that is good though :smile: