@Lucas-C 2023-08-16
The only remaining failing test is `tests/test_writer.py::test_iss1601`!