@Lucas-C 2023-08-20
There are 2 tests that remain to fail:
```
FAILED tests/test_workflows.py::test_extra_test_iss1541 - Failed: DID NOT RAISE <class 'pypdf.errors.PdfReadError'>
FAILED tests/test_writer.py::test_iss1601 - AssertionError: assert b'q\nq 1 0 0 1 0 0 cm ...\n'
```